# power-platform/openapi-specification.json
{
  "swagger": "2.0",
  "info": {
    "title": "Designetica AI Wireframe Connector",
    "description": "Advanced AI-powered wireframe generation with image analysis, component detection, and Power Platform integration",
    "version": "2.0.0",
    "contact": {
      "name": "Designetica API Support",
      "email": "support@designetica.com"
    }
  },
  "host": "func-designetica-prod-xabnur6oyusju.azurewebsites.net",
  "basePath": "/api",
  "schemes": ["https"],
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "oauth2_auth": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://login.microsoftonline.com/common/oauth2/authorize",
      "tokenUrl": "https://login.microsoftonline.com/common/oauth2/token",
      "scopes": {
        "https://graph.microsoft.com/.default": "Access Microsoft Graph"
      }
    }
  },
  "security": [{ "oauth2_auth": ["https://graph.microsoft.com/.default"] }],
  "paths": {
    "/generate-wireframe": {
      "post": {
        "summary": "Generate AI-powered wireframe from description",
        "description": "Creates professional wireframes using GPT-4o with context-aware generation",
        "operationId": "GenerateWireframe",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "type": "string",
                  "description": "Wireframe description",
                  "x-ms-summary": "Description"
                },
                "colorScheme": {
                  "type": "string",
                  "enum": ["primary", "secondary", "dark", "light"],
                  "default": "primary",
                  "description": "Color scheme to use",
                  "x-ms-summary": "Color Scheme"
                },
                "sessionId": {
                  "type": "string",
                  "description": "Session ID for context awareness",
                  "x-ms-summary": "Session ID"
                },
                "imageAnalysis": {
                  "type": "object",
                  "description": "Image analysis data from AI Builder",
                  "x-ms-summary": "Image Analysis"
                },
                "useAI": {
                  "type": "boolean",
                  "default": true,
                  "description": "Use AI generation",
                  "x-ms-summary": "Use AI"
                }
              },
              "required": ["description"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wireframe generated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "html": {
                  "type": "string",
                  "description": "Generated wireframe HTML"
                },
                "aiGenerated": {
                  "type": "boolean",
                  "description": "Whether AI was used"
                },
                "source": {
                  "type": "string",
                  "description": "Generation source"
                },
                "processingTimeMs": {
                  "type": "number",
                  "description": "Processing time in milliseconds"
                },
                "correlationId": {
                  "type": "string",
                  "description": "Request correlation ID"
                },
                "confidence": {
                  "type": "number",
                  "description": "AI confidence score"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}